factorized ReactionsListWithInformation
